refactor: extract git operation helper modules
