+++ src/core/git/operations.rs
@@ -4,29 +4,11 @@ use std::process::Command;
 
 use crate::core::config::read_json_spec_to_string;
 use crate::core::error::{Error, Result};
-use crate::core::output::{BulkResult, BulkSummary, ItemOutcome};
-use crate::core::project;
-use crate::core::release::changelog;
+use crate::core::output::BulkResult;
 
-use super::changes::*;
-use super::commits::*;
 use super::operation_output::{run_bulk_ids, GitOutput};
 use super::primitives::is_git_repo;
 use super::{execute_git, resolve_target};
-
-const DEFAULT_COMMIT_LIMIT: usize = 10;
-const VERBOSE_UNTRACKED_THRESHOLD: usize = 200;
-
-const NOISY_UNTRACKED_DIRS: [&str; 8] = [
-    "node_modules",
-    "dist",
-    "build",
-    "coverage",
-    ".next",
-    "vendor",
-    "target",
-    ".cache",
-];
 
 #[derive(Debug, Clone, Serialize)]
 pub struct RepoSnapshot {
@@ -38,261 +20,11 @@ pub struct RepoSnapshot {
     pub behind: Option<u32>,
 }
 
-#[derive(Debug, Clone, Serialize)]
-pub struct RepoBaselineSnapshot {
-    pub branch: String,
-    pub clean: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ahead: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub behind: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub commits_since_version: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub baseline_ref: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub baseline_warning: Option<String>,
-}
-
-// === Changes Output Types ===
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum BaselineSource {
-    Tag,
-    VersionCommit,
-    LastNCommits,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct ChangelogInfo {
-    pub unreleased_entries: usize,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub path: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub hint: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-
-pub struct ChangesOutput {
-    pub component_id: String,
-    pub path: String,
-    pub success: bool,
-    pub latest_tag: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub baseline_source: Option<BaselineSource>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub baseline_ref: Option<String>,
-    pub commits: Vec<CommitInfo>,
-    pub uncommitted: UncommittedChanges,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub uncommitted_diff: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub diff: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub warning: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub changelog: Option<ChangelogInfo>,
-}
-
-pub struct BaselineInfo {
-    pub latest_tag: Option<String>,
-    pub source: Option<BaselineSource>,
-    pub reference: Option<String>,
-    pub warning: Option<String>,
-}
-
-pub fn build_repo_baseline_snapshot(
-    path: &str,
-    current_version: Option<&str>,
-) -> Result<RepoBaselineSnapshot> {
-    let snapshot = get_repo_snapshot(path)?;
-    let baseline = detect_baseline_with_version(path, current_version).ok();
-    let commits_since = baseline.as_ref().and_then(|b| {
-        get_commits_since_tag(path, b.reference.as_deref())
-            .ok()
-            .map(|c| c.len() as u32)
-    });
-
-    Ok(RepoBaselineSnapshot {
-        branch: snapshot.branch,
-        clean: snapshot.clean,
-        ahead: snapshot.ahead,
-        behind: snapshot.behind,
-        commits_since_version: commits_since,
-        baseline_ref: baseline.as_ref().and_then(|b| b.reference.clone()),
-        baseline_warning: baseline.and_then(|b| b.warning),
-    })
-}
-
-/// Detect baseline with version alignment checking.
-/// If a tag exists but doesn't match current_version, warns and finds version commit instead.
-pub fn detect_baseline_with_version(
-    path: &str,
-    current_version: Option<&str>,
-) -> Result<BaselineInfo> {
-    // Fetch tags from remote so locally-missing tags (pushed from another
-    // machine) are available before we resolve the baseline.  Best-effort:
-    // if there is no remote or the network is unavailable we silently
-    // proceed with whatever tags are already local.
-    let _ =
-        crate::core::engine::command::run_in_optional(path, "git", &["fetch", "--tags", "--quiet"]);
-
-    // Priority 1: Check for latest tag
-    if let Some(tag) = get_latest_tag(path)? {
-        let tag_version = extract_version_from_tag(&tag);
-
-        // If we have current version, check alignment
-        if let (Some(current), Some(tag_ver)) = (current_version, &tag_version) {
-            if current != tag_ver {
-                // Tag is stale - try to find the release commit for current version
-                if let Some(hash) = find_version_release_commit(path, current)? {
-                    return Ok(BaselineInfo {
-                        latest_tag: Some(tag.clone()),
-                        source: Some(BaselineSource::VersionCommit),
-                        reference: Some(hash),
-                        warning: Some(format!(
-                            "Latest tag '{}' doesn't match version {}. Using release commit as baseline. Consider: git tag v{}",
-                            tag, current, current
-                        )),
-                    });
-                }
-
-                // No matching release commit - fall back to generic version commit
-                if let Some(hash) = find_version_commit(path)? {
-                    return Ok(version_commit_baseline(
-                        Some(tag.clone()),
-                        hash,
-                        format!(
-                            "Latest tag '{}' doesn't match version {}. Using most recent version commit.",
-                            tag, current
-                        ),
-                    ));
-                }
-
-                // No version commits found - use the stale tag but warn
-                return Ok(BaselineInfo {
-                    latest_tag: Some(tag.clone()),
-                    source: Some(BaselineSource::Tag),
-                    reference: Some(tag.clone()),
-                    warning: Some(format!(
-                        "Latest tag '{}' doesn't match version {}. Consider: git tag v{}",
-                        tag, current, current
-                    )),
-                });
-            }
-        }
-
-        // Tag version matches or no version to compare - use tag
-        return Ok(BaselineInfo {
-            latest_tag: Some(tag.clone()),
-            source: Some(BaselineSource::Tag),
-            reference: Some(tag),
-            warning: None,
-        });
-    }
-
-    // Priority 2: No tags - try version commit for current version first
-    if let Some(current) = current_version {
-        if let Some(hash) = find_version_release_commit(path, current)? {
-            return Ok(BaselineInfo {
-                latest_tag: None,
-                source: Some(BaselineSource::VersionCommit),
-                reference: Some(hash),
-                warning: Some(
-                    "No tags found. Using release commit for current version.".to_string(),
-                ),
-            });
-        }
-    }
-
-    // Priority 3: Generic version commit
-    if let Some(hash) = find_version_commit(path)? {
-        return Ok(version_commit_baseline(
-            None,
-            hash,
-            "No tags found. Using most recent version commit as baseline.".to_string(),
-        ));
-    }
-
-    // Fallback: No baseline found
-    Ok(BaselineInfo {
-        latest_tag: None,
-        source: Some(BaselineSource::LastNCommits),
-        reference: None,
-        warning: Some(format!(
-            "No tags or version commits found. Showing last {} commits.",
-            DEFAULT_COMMIT_LIMIT
-        )),
-    })
-}
-
-fn version_commit_baseline(
-    latest_tag: Option<String>,
-    reference: String,
-    warning: String,
-) -> BaselineInfo {
-    BaselineInfo {
-        latest_tag,
-        source: Some(BaselineSource::VersionCommit),
-        reference: Some(reference),
-        warning: Some(warning),
-    }
-}
-
 // Input types for JSON parsing
 #[derive(Debug, Deserialize)]
 
-struct BulkIdsInput {
-    component_ids: Vec<String>,
-    #[serde(default)]
-    tags: bool,
-    #[serde(default)]
-    force_with_lease: bool,
-    #[serde(default)]
-    remote_url: Option<String>,
-    #[serde(default)]
-    token: Option<String>,
-    #[serde(default)]
-    refspec: Option<String>,
-    #[serde(default)]
-    strip_extraheader: bool,
-}
-
-#[derive(Debug, Deserialize)]
-
-struct BulkCommitInput {
-    components: Vec<CommitSpec>,
-}
-
-#[derive(Debug, Deserialize)]
-
-struct CommitSpec {
-    #[serde(default)]
-    id: Option<String>,
-    message: String,
-    #[serde(default)]
-    staged_only: bool,
-    #[serde(default, alias = "include_files")]
-    files: Option<Vec<String>>,
-    #[serde(default, alias = "exclude_files")]
-    exclude_files: Option<Vec<String>>,
-}
-
-/// Options for commit operations.
-#[derive(Debug, Clone, Default)]
-pub struct CommitOptions {
-    /// Skip `git add` and commit only staged changes
-    pub staged_only: bool,
-    /// Stage and commit only these specific files
-    pub files: Option<Vec<String>>,
-    /// Stage all except these files (mutually exclusive with `files`)
-    pub exclude: Option<Vec<String>>,
-    /// Amend the previous commit instead of creating a new one
-    pub amend: bool,
+pub(crate) struct BulkIdsInput {
+    pub(crate) component_ids: Vec<String>,
 }
 
 pub fn execute_git_for_release(path: &str, args: &[&str]) -> std::io::Result<std::process::Output> {
@@ -356,71 +88,6 @@ fn parse_ahead_behind(counts: &str) -> (Option<u32>, Option<u32>) {
     (ahead, behind)
 }
 
-pub(crate) fn build_untracked_hint(path: &str, untracked_count: usize) -> Option<String> {
-    if untracked_count < VERBOSE_UNTRACKED_THRESHOLD {
-        return None;
-    }
-
-    let ignored_output = execute_git(path, &["status", "--ignored", "--porcelain=v1"]).ok()?;
-    if !ignored_output.status.success() {
-        return None;
-    }
-
-    let stdout = String::from_utf8_lossy(&ignored_output.stdout);
-    let ignored_lines: Vec<&str> = stdout
-        .lines()
-        .filter(|line| line.starts_with("!!"))
-        .collect();
-    if ignored_lines.is_empty() {
-        return None;
-    }
-
-    let mut noisy_ignored = Vec::new();
-    for line in ignored_lines {
-        let path = line[3..].trim();
-        for dir in NOISY_UNTRACKED_DIRS {
-            if path == dir || path.starts_with(&format!("{}/", dir)) {
-                noisy_ignored.push(dir.to_string());
-                break;
-            }
-        }
-    }
-
-    noisy_ignored.sort();
-    noisy_ignored.dedup();
-
-    if noisy_ignored.is_empty() {
-        return None;
-    }
-
-    Some(format!(
-        "Large untracked list detected ({}). Common noisy directories ignored by git: {}. If output feels too big, add them to .gitignore.",
-        untracked_count,
-        noisy_ignored.join(", ")
-    ))
-}
-
-fn resolve_changelog_info(
-    component: &crate::core::component::Component,
-    _commits: &[CommitInfo],
-) -> Option<ChangelogInfo> {
-    let changelog_path = changelog::resolve_changelog_path(component).ok()?;
-    let content = std::fs::read_to_string(&changelog_path).ok()?;
-    let settings = changelog::resolve_effective_settings(Some(component));
-    let unreleased_entries =
-        changelog::count_unreleased_entries(&content, &settings.next_section_aliases);
-
-    // No hint: homeboy auto-generates changelog entries from commits at
-    // release time, so an empty `## Unreleased` section no longer implies
-    // the user needs to do anything. The count itself is still useful
-    // context for `homeboy changes` output.
-    Some(ChangelogInfo {
-        unreleased_entries,
-        path: Some(changelog_path.to_string_lossy().to_string()),
-        hint: None,
-    })
-}
-
 /// Get git status for a component.
 pub fn status(component_id: Option<&str>) -> Result<GitOutput> {
     status_at(component_id, None)
@@ -446,365 +113,6 @@ pub fn status_bulk(json_spec: &str) -> Result<BulkResult<GitOutput>> {
     })?;
     Ok(run_bulk_ids(&input.component_ids, "status", |id| {
         status(Some(id))
-    }))
-}
-
-/// Commit changes for a component.
-///
-/// By default, stages all changes before committing. Use `options` to control staging:
-/// - `staged_only`: Skip staging, commit only what's already staged
-/// - `files`: Stage only these specific files before committing
-///
-/// When `path_override` is provided, git operations run in that directory
-/// instead of the component's configured `local_path`.
-pub fn commit(
-    component_id: Option<&str>,
-    message: Option<&str>,
-    options: CommitOptions,
-) -> Result<GitOutput> {
-    commit_at(component_id, message, options, None)
-}
-
-/// Like [`commit`] but with an explicit path override for git operations.
-pub fn commit_at(
-    component_id: Option<&str>,
-    message: Option<&str>,
-    options: CommitOptions,
-    path_override: Option<&str>,
-) -> Result<GitOutput> {
-    let msg = message.ok_or_else(|| {
-        Error::validation_invalid_argument("message", "Missing commit message", None, None)
-    })?;
-    let (id, path) = resolve_target(component_id, path_override)?;
-
-    // Check for changes - behavior differs based on staged_only
-    let status_output = execute_git(&path, &["status", "--porcelain=v1"])
-        .map_err(|e| Error::git_command_failed(e.to_string()))?;
-    let status_str = String::from_utf8_lossy(&status_output.stdout);
-
-    if options.staged_only {
-        // Check if there are staged changes (lines starting with non-space in first column)
-        let has_staged = status_str.lines().any(|line| {
-            let first_char = line.chars().next().unwrap_or(' ');
-            first_char != ' ' && first_char != '?'
-        });
-        if !has_staged {
-            return Ok(GitOutput {
-                component_id: id,
-                path,
-                action: "commit".to_string(),
-                success: true,
-                exit_code: 0,
-                stdout: "Nothing staged to commit".to_string(),
-                stderr: String::new(),
-            });
-        }
-    } else if status_str.trim().is_empty() {
-        return Ok(GitOutput {
-            component_id: id,
-            path,
-            action: "commit".to_string(),
-            success: true,
-            exit_code: 0,
-            stdout: "Nothing to commit, working tree clean".to_string(),
-            stderr: String::new(),
-        });
-    }
-
-    // Stage changes based on options
-    if !options.staged_only {
-        match (&options.files, &options.exclude) {
-            // Both specified: error (mutually exclusive)
-            (Some(_), Some(_)) => {
-                return Err(Error::validation_invalid_argument(
-                    "files/exclude",
-                    "Cannot use both --files and --exclude",
-                    None,
-                    None,
-                ));
-            }
-            // Include only specific files
-            (Some(files), None) => {
-                let mut args = vec!["add", "--"];
-                args.extend(files.iter().map(|s| s.as_str()));
-                let add_output = execute_git(&path, &args)
-                    .map_err(|e| Error::git_command_failed(e.to_string()))?;
-                if !add_output.status.success() {
-                    return Ok(GitOutput::from_output(id, path, "commit", add_output));
-                }
-            }
-            // Exclude specific files: stage all, then unstage excluded
-            (None, Some(excluded)) => {
-                let add_output = execute_git(&path, &["add", "."])
-                    .map_err(|e| Error::git_command_failed(e.to_string()))?;
-                if !add_output.status.success() {
-                    return Ok(GitOutput::from_output(id, path, "commit", add_output));
-                }
-                // Unstage excluded files (git reset -- file1 file2)
-                // Note: git reset without --hard only unstages, does not discard changes
-                let mut reset_args = vec!["reset", "--"];
-                reset_args.extend(excluded.iter().map(|s| s.as_str()));
-                let reset_output = execute_git(&path, &reset_args)
-                    .map_err(|e| Error::git_command_failed(e.to_string()))?;
-                if !reset_output.status.success() {
-                    return Ok(GitOutput::from_output(id, path, "commit", reset_output));
-                }
-            }
-            // Default: stage all
-            (None, None) => {
-                let add_output = execute_git(&path, &["add", "."])
-                    .map_err(|e| Error::git_command_failed(e.to_string()))?;
-                if !add_output.status.success() {
-                    return Ok(GitOutput::from_output(id, path, "commit", add_output));
-                }
-            }
-        }
-    }
-
-    let args: Vec<&str> = if options.amend {
-        vec!["commit", "--amend", "-m", msg]
-    } else {
-        vec!["commit", "-m", msg]
-    };
-    let output = execute_git(&path, &args).map_err(|e| Error::git_command_failed(e.to_string()))?;
-    Ok(GitOutput::from_output(id, path, "commit", output))
-}
-
-/// Commit multiple components from JSON spec (bulk format).
-fn commit_bulk(json_spec: &str) -> Result<BulkResult<GitOutput>> {
-    let input: BulkCommitInput = serde_json::from_str(json_spec).map_err(|e| {
-        Error::validation_invalid_json(
-            e,
-            Some("parse bulk commit input".to_string()),
-            Some(json_spec.chars().take(200).collect::<String>()),
-        )
-    })?;
-
-    let mut results = Vec::new();
-    let mut succeeded = 0usize;
-    let mut failed = 0usize;
-
-    for spec in &input.components {
-        let id = match &spec.id {
-            Some(id) => id.clone(),
-            None => {
-                failed += 1;
-                results.push(ItemOutcome {
-                    id: "unknown".to_string(),
-                    result: None,
-                    error: Some("Missing 'id' field in bulk commit spec".to_string()),
-                });
-                continue;
-            }
-        };
-        let options = CommitOptions {
-            staged_only: spec.staged_only,
-            files: spec.files.clone(),
-            exclude: spec.exclude_files.clone(),
-            amend: false,
-        };
-        match commit(Some(&id), Some(&spec.message), options) {
-            Ok(output) => {
-                if output.success {
-                    succeeded += 1;
-                } else {
-                    failed += 1;
-                }
-                results.push(ItemOutcome {
-                    id,
-                    result: Some(output),
-                    error: None,
-                });
-            }
-            Err(e) => {
-                failed += 1;
-                results.push(ItemOutcome {
-                    id,
-                    result: None,
-                    error: Some(e.to_string()),
-                });
-            }
-        }
-    }
-
-    Ok(BulkResult {
-        action: "commit".to_string(),
-        results,
-        summary: BulkSummary {
-            total: succeeded + failed,
-            succeeded,
-            failed,
-        },
-    })
-}
-
-/// Output from commit_from_json - either single or bulk result.
-#[derive(Serialize)]
-#[serde(untagged)]
-pub enum CommitJsonOutput {
-    Single(GitOutput),
-    Bulk(BulkResult<GitOutput>),
-}
-
-/// Commit from JSON spec. Auto-detects single vs bulk format.
-///
-/// Single format: `{"id":"x","message":"m"}` or `{"message":"m"}` (uses CWD/positional ID)
-/// Bulk format: `{"components":[{"id":"x","message":"m"},...]}`
-pub fn commit_from_json(id: Option<&str>, json_spec: &str) -> Result<CommitJsonOutput> {
-    let raw = read_json_spec_to_string(json_spec)?;
-    let parsed: serde_json::Value = serde_json::from_str(&raw).map_err(|e| {
-        Error::validation_invalid_json(
-            e,
-            Some("parse commit json".to_string()),
-            Some(raw.chars().take(200).collect::<String>()),
-        )
-    })?;
-
-    // Auto-detect: bulk if has "components" array
-    if parsed.get("components").is_some() {
-        let bulk = commit_bulk(&raw)?;
-        return Ok(CommitJsonOutput::Bulk(bulk));
-    }
-
-    // Single spec - parse and extract fields
-    let spec: CommitSpec = serde_json::from_str(&raw).map_err(|e| {
-        Error::validation_invalid_json(
-            e,
-            Some("parse commit spec".to_string()),
-            Some(raw.chars().take(200).collect::<String>()),
-        )
-    })?;
-
-    // ID priority: positional arg > JSON body
-    let target_id = id.map(|s| s.to_string()).or(spec.id);
-    let options = CommitOptions {
-        staged_only: spec.staged_only,
-        files: spec.files,
-        exclude: spec.exclude_files,
-        amend: false,
-    };
-
-    let output = commit(target_id.as_deref(), Some(&spec.message), options)?;
-    Ok(CommitJsonOutput::Single(output))
-}
-
-/// Options for [`push`].
-#[derive(Debug, Clone, Default)]
-pub struct PushOptions {
-    /// Push tags as well (`--follow-tags`).
-    pub tags: bool,
-    /// Use `--force-with-lease` for safe force-pushes (e.g. after a rebase).
-    /// Deliberately the only force flavour exposed — never plain `--force`.
-    pub force_with_lease: bool,
-    /// Push to a remote URL directly instead of the configured upstream.
-    pub remote_url: Option<String>,
-    /// GitHub App/user token injected into `remote_url` for this invocation.
-    pub token: Option<String>,
-    /// Explicit source/destination refspec, e.g. `HEAD:refs/heads/branch`.
-    pub refspec: Option<String>,
-    /// Clear the GitHub Actions checkout extraheader so URL auth wins.
-    pub strip_extraheader: bool,
-}
-
-/// Push local commits for a component.
-pub fn push(component_id: Option<&str>, options: PushOptions) -> Result<GitOutput> {
-    push_at(component_id, options, None)
-}
-
-/// Like [`push`] but with an explicit path override for git operations.
-pub fn push_at(
-    component_id: Option<&str>,
-    options: PushOptions,
-    path_override: Option<&str>,
-) -> Result<GitOutput> {
-    let (id, path) = resolve_target(component_id, path_override)?;
-    let remote_url =
-        resolve_push_remote_url(options.remote_url.as_deref(), options.token.as_deref())?;
-    let mut args: Vec<String> = Vec::new();
-    if options.strip_extraheader {
-        args.push("-c".to_string());
-        args.push("http.https://github.com/.extraheader=".to_string());
-    }
-    args.push("push".to_string());
-    if options.tags {
-        args.push("--follow-tags".to_string());
-    }
-    if options.force_with_lease {
-        args.push("--force-with-lease".to_string());
-    }
-    if let Some(remote) = remote_url {
-        args.push(remote);
-    } else if options.refspec.is_some() {
-        args.push("origin".to_string());
-    }
-    if let Some(refspec) = options.refspec {
-        args.push(refspec);
-    }
-    let arg_refs: Vec<&str> = args.iter().map(String::as_str).collect();
-    let output =
-        execute_git(&path, &arg_refs).map_err(|e| Error::git_command_failed(e.to_string()))?;
-    Ok(GitOutput::from_output(id, path, "push", output))
-}
-
-fn resolve_push_remote_url(
-    remote_url: Option<&str>,
-    token: Option<&str>,
-) -> Result<Option<String>> {
-    match (remote_url, token) {
-        (Some(url), Some(token)) => {
-            if !url.starts_with("https://github.com/") {
-                return Err(Error::validation_invalid_argument(
-                    "token",
-                    "--token requires --remote-url to start with https://github.com/",
-                    None,
-                    None,
-                ));
-            }
-            Ok(Some(format!(
-                "https://x-access-token:{}@{}",
-                token,
-                &url["https://".len()..]
-            )))
-        }
-        (Some(url), None) => Ok(Some(url.to_string())),
-        (None, Some(_)) => Err(Error::validation_invalid_argument(
-            "token",
-            "--token requires --remote-url",
-            None,
-            None,
-        )),
-        (None, None) => Ok(None),
-    }
-}
-
-/// Push multiple components from JSON spec.
-pub fn push_bulk(json_spec: &str) -> Result<BulkResult<GitOutput>> {
-    let raw = read_json_spec_to_string(json_spec)?;
-    let input: BulkIdsInput = serde_json::from_str(&raw).map_err(|e| {
-        Error::validation_invalid_json(
-            e,
-            Some("parse bulk push input".to_string()),
-            Some(raw.chars().take(200).collect::<String>()),
-        )
-    })?;
-    let push_tags = input.tags;
-    let force_with_lease = input.force_with_lease;
-    let remote_url = input.remote_url;
-    let token = input.token;
-    let refspec = input.refspec;
-    let strip_extraheader = input.strip_extraheader;
-    Ok(run_bulk_ids(&input.component_ids, "push", |id| {
-        push(
-            Some(id),
-            PushOptions {
-                tags: push_tags,
-                force_with_lease,
-                remote_url: remote_url.clone(),
-                token: token.clone(),
-                refspec: refspec.clone(),
-                strip_extraheader,
-            },
-        )
     }))
 }
 
@@ -1180,190 +488,4 @@ pub fn fetch_and_fast_forward(path: &str) -> Result<Option<u32>> {
             Ok(Some(n))
         }
     }
-}
-
-/// Get all changes for a component.
-pub fn changes(
-    component_id: Option<&str>,
-    since_tag: Option<&str>,
-    include_diff: bool,
-) -> Result<ChangesOutput> {
-    changes_at(component_id, since_tag, include_diff, None)
-}
-
-/// Like [`changes`] but with an explicit path override for git operations.
-pub fn changes_at(
-    component_id: Option<&str>,
-    since_tag: Option<&str>,
-    include_diff: bool,
-    path_override: Option<&str>,
-) -> Result<ChangesOutput> {
-    let (id, path) = resolve_target(component_id, path_override)?;
-
-    // Load component for version checking and changelog info
-    let component = crate::core::component::resolve_effective(Some(&id), Some(&path), None).ok();
-
-    // Determine baseline with version alignment awareness
-    let baseline = match since_tag {
-        Some(t) => {
-            // Explicit tag override - use as-is
-            BaselineInfo {
-                latest_tag: Some(t.to_string()),
-                source: Some(BaselineSource::Tag),
-                reference: Some(t.to_string()),
-                warning: None,
-            }
-        }
-        None => {
-            // Use component version for alignment checking
-            let current_version = component
-                .as_ref()
-                .and_then(crate::core::release::version::get_component_version);
-            detect_baseline_with_version(&path, current_version.as_deref())?
-        }
-    };
-
-    let commits = match baseline.source {
-        Some(BaselineSource::LastNCommits) => get_last_n_commits(&path, DEFAULT_COMMIT_LIMIT)?,
-        _ => get_commits_since_tag(&path, baseline.reference.as_deref())?,
-    };
-
-    // Resolve changelog info if component has changelog configured
-    let changelog_info = component
-        .as_ref()
-        .and_then(|c| resolve_changelog_info(c, &commits));
-
-    let uncommitted = get_uncommitted_changes(&path)?;
-    let uncommitted_diff = if uncommitted.has_changes {
-        Some(get_diff(&path)?)
-    } else {
-        None
-    };
-    let diff = if include_diff {
-        baseline
-            .reference
-            .as_ref()
-            .map(|r| get_range_diff(&path, r))
-            .transpose()?
-    } else {
-        None
-    };
-
-    Ok(ChangesOutput {
-        component_id: id,
-        path,
-        success: true,
-        latest_tag: baseline.latest_tag,
-        baseline_source: baseline.source,
-        baseline_ref: baseline.reference,
-        commits,
-        uncommitted,
-        uncommitted_diff,
-        diff,
-        warning: baseline.warning,
-        error: None,
-        changelog: changelog_info,
-    })
-}
-
-fn build_bulk_changes_output(
-    component_ids: &[String],
-    include_diff: bool,
-) -> BulkResult<ChangesOutput> {
-    let mut results = Vec::new();
-    let mut succeeded = 0usize;
-    let mut failed = 0usize;
-
-    for id in component_ids {
-        match changes(Some(id), None, include_diff) {
-            Ok(output) => {
-                if output.success {
-                    succeeded += 1;
-                } else {
-                    failed += 1;
-                }
-                results.push(ItemOutcome {
-                    id: id.clone(),
-                    result: Some(output),
-                    error: None,
-                });
-            }
-            Err(e) => {
-                failed += 1;
-                results.push(ItemOutcome {
-                    id: id.clone(),
-                    result: None,
-                    error: Some(e.to_string()),
-                });
-            }
-        }
-    }
-
-    BulkResult {
-        action: "changes".to_string(),
-        results,
-        summary: BulkSummary {
-            total: succeeded + failed,
-            succeeded,
-            failed,
-        },
-    }
-}
-
-/// Get changes for multiple components from JSON spec.
-pub fn changes_bulk(json_spec: &str, include_diff: bool) -> Result<BulkResult<ChangesOutput>> {
-    let raw = read_json_spec_to_string(json_spec)?;
-    let input: BulkIdsInput = serde_json::from_str(&raw).map_err(|e| {
-        Error::validation_invalid_json(
-            e,
-            Some("parse bulk changes input".to_string()),
-            Some(raw.chars().take(200).collect::<String>()),
-        )
-    })?;
-
-    Ok(build_bulk_changes_output(
-        &input.component_ids,
-        include_diff,
-    ))
-}
-
-/// Get changes for all components in a project.
-pub fn changes_project(project_id: &str, include_diff: bool) -> Result<BulkResult<ChangesOutput>> {
-    let proj = project::load(project_id)?;
-    let component_ids: Vec<String> = project::resolve_project_components(&proj)?
-        .into_iter()
-        .map(|component| component.id)
-        .collect();
-    Ok(build_bulk_changes_output(&component_ids, include_diff))
-}
-
-/// Get changes for specific components in a project (filtered).
-pub fn changes_project_filtered(
-    project_id: &str,
-    component_ids: &[String],
-    include_diff: bool,
-) -> Result<BulkResult<ChangesOutput>> {
-    let proj = project::load(project_id)?;
-
-    // Filter to only components that are in the project
-    let filtered: Vec<String> = component_ids
-        .iter()
-        .filter(|id| project::has_component(&proj, id))
-        .cloned()
-        .collect();
-
-    if filtered.is_empty() {
-        return Err(Error::validation_invalid_argument(
-            "component_ids",
-            format!(
-                "None of the specified components are in project '{}'. Available: {}",
-                project_id,
-                project::project_component_ids(&proj).join(", ")
-            ),
-            None,
-            None,
-        ));
-    }
-
-    Ok(build_bulk_changes_output(&filtered, include_diff))
 }
